backend/requirements.txt
# =============================================================================
# requirements.txt
# =============================================================================
Flask
Flask-PyMongo
Flask-JWT-Extended
bcrypt
python-dotenv
pymongo
bson